backend/requirements.txt
fastapi
fastmcp
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib
mcp
openai
requests
python-dotenv
tzdata
psycopg2-binary
alembic